alexa_skill/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
Django==2.1.2
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
pipenv==2018.7.1
pymongo==3.7.1
pytz==2018.5
requests==2.19.1
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
